+++ src/downloader/__init__.py
@@ -1,5 +1,5 @@
 """Downloader module for OpenDPM."""
 
-from .download import download_source, extract_archive
+from .fetcher import download_source, extract_archive
 
 __all__ = ["download_source", "extract_archive"]
